+++ MelonLoader.Support.Mono.Pre2017/Main.cs
@@ -38,11 +38,11 @@ namespace MelonLoader.Support
         }
         internal static void Destroy() { Main.IsDestroying = true; if (Main.obj != null) GameObject.Destroy(Main.obj); }
         void Start() => transform.SetAsLastSibling();
-        void Update() { transform.SetAsLastSibling(); MelonLoader.Main.OnUpdate(); }
-        void FixedUpdate() => MelonLoader.Main.OnFixedUpdate();
-        void LateUpdate() => MelonLoader.Main.OnLateUpdate();
-        void OnGUI() => MelonLoader.Main.OnGUI();
+        void Update() { transform.SetAsLastSibling(); MelonHandler.OnUpdate(); }
+        void FixedUpdate() => MelonHandler.OnFixedUpdate();
+        void LateUpdate() => MelonHandler.OnLateUpdate();
+        void OnGUI() => MelonHandler.OnGUI();
         void OnDestroy() { if (!Main.IsDestroying) Create(); }
-        void OnApplicationQuit() { Destroy(); MelonLoader.Main.OnApplicationQuit(); }
+        void OnApplicationQuit() { Destroy(); MelonHandler.OnApplicationQuit(); }
     }
 }