+++ MelonLoader.Support.Mono/Main.cs
@@ -44,11 +44,11 @@ namespace MelonLoader.Support
         }
 
         void Start() => transform.SetAsLastSibling();
-        void Update() { transform.SetAsLastSibling(); MelonLoader.Main.OnUpdate(); }
-        void FixedUpdate() => MelonLoader.Main.OnFixedUpdate();
-        void LateUpdate() => MelonLoader.Main.OnLateUpdate();
-        void OnGUI() => MelonLoader.Main.OnGUI();
+        void Update() { transform.SetAsLastSibling(); MelonHandler.OnUpdate(); }
+        void FixedUpdate() => MelonHandler.OnFixedUpdate();
+        void LateUpdate() => MelonHandler.OnLateUpdate();
+        void OnGUI() => MelonHandler.OnGUI();
         void OnDestroy() { if (!Main.IsDestroying) Create(); }
-        void OnApplicationQuit() { Destroy(); MelonLoader.Main.OnApplicationQuit(); }
+        void OnApplicationQuit() { Destroy(); MelonHandler.OnApplicationQuit(); }
     }
 }